Removed state classes and added more tests for storage
