+++ modelrunner/storage/backend/zarr.py
@@ -146,7 +146,7 @@ class ZarrStorage(StorageBase):
     def _read_attrs(self, loc: Sequence[str]) -> AttrsLike:
         return self[loc].attrs  # type: ignore
 
-    def _write_attr(self, loc: Sequence[str], name: str, value) -> None:
+    def _write_attr(self, loc: Sequence[str], name: str, value: str) -> None:
         self[loc].attrs[name] = value
 
     def _read_array(
@@ -209,6 +209,7 @@ class ZarrStorage(StorageBase):
         return self[loc][0]
 
     def _write_object(self, loc: Sequence[str], obj: Any) -> None:
-        arr = np.array([obj], dtype=object)  # encode object in an array
+        arr = np.empty(1, dtype=object)  # encode object in an array
+        arr[0] = obj
         parent, name = self._get_parent(loc)
         parent.array(name, arr, object_codec=self.codec, overwrite=True)
