+++ tests/compatibility/2/array.json
@@ -1,1 +1,1 @@
-{"result": {"__attrs__": {"__model__": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "__version__": "2", "__class__": "\"modelrunner.results.Result\""}, "state": {"__type__": "array", "data": [0, 1, 2], "dtype": "{\"__pickled__\": \"gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAmk4lImIh5RSlChLA4wBPJROTk5K/////0r/\\n////SwB0lGIu\\n\"}", "__attrs__": {"__class__": "\"modelrunner.state.array.ArrayState\"", "__version__": "2"}}}}
+{"__attrs__": {"__codec__": {"id": "pickle", "protocol": 5}}, "result": {"data": {"__pickled__": "gASVmgAAAAAAAABDloAFlYsAAAAAAAAAjBJudW1weS5jb3JlLm51bWVyaWOUjAtfZnJvbWJ1ZmZl\ncpSTlCiWGAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAgAAAAAAAACUjAVudW1weZSMBWR0eXBllJOU\njAJpOJSJiIeUUpQoSwOMATyUTk5OSv////9K/////0sAdJRiSwOFlIwBQ5R0lFKULpQu\n"}, "__attrs__": {"model": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "format_version": "2", "__type__": "object", "__class__": "modelrunner.results.Result"}}}
