+++ tests/compatibility/2/array_col.json
@@ -1,1 +1,1 @@
-{"result": {"__attrs__": {"__model__": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "__version__": "2", "__class__": "\"modelrunner.results.Result\""}, "state": {"__attrs__": {"labels": "[\"a\", \"b\"]", "__class__": "\"modelrunner.state.array_collection.ArrayCollectionState\"", "__version__": "2"}, "a": {"__type__": "array", "data": [0, 1], "dtype": "{\"__pickled__\": \"gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAmk4lImIh5RSlChLA4wBPJROTk5K/////0r/\\n////SwB0lGIu\\n\"}"}, "b": {"__type__": "array", "data": [0, 1, 2], "dtype": "{\"__pickled__\": \"gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAmk4lImIh5RSlChLA4wBPJROTk5K/////0r/\\n////SwB0lGIu\\n\"}"}}}}
+{"__attrs__": {"__codec__": {"id": "pickle", "protocol": 5}}, "result": {"data": {"__pickled__": "gASVxwAAAAAAAABDw4AFlbgAAAAAAAAAXZQojBJudW1weS5jb3JlLm51bWVyaWOUjAtfZnJvbWJ1\nZmZlcpSTlCiWEAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAlIwFbnVtcHmUjAVkdHlwZZSTlIwCaTiU\niYiHlFKUKEsDjAE8lE5OTkr/////Sv////9LAHSUYksChZSMAUOUdJRSlGgDKJYYAAAAAAAAAAAA\nAAAAAAAAAQAAAAAAAAACAAAAAAAAAJRoCksDhZRoDnSUUpRlLpQu\n"}, "__attrs__": {"model": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "format_version": "2", "__type__": "object", "__class__": "modelrunner.results.Result"}}}
