+++ tests/compatibility/2/dict.json
@@ -1,1 +1,1 @@
-{"result": {"__attrs__": {"__model__": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "__version__": "2", "__class__": "\"modelrunner.results.Result\""}, "state": {"__attrs__": {"__keys__": "[\"a\", \"b\"]", "__class__": "\"modelrunner.state.dict.DictState\"", "__version__": "2"}, "a": {"__type__": "array", "data": [["a", "b"]], "dtype": "{\"__pickled__\": \"gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAk84lImIh5RSlChLA4wBfJROTk5K/////0r/\\n////Sz90lGIu\\n\"}", "__attrs__": {"__class__": "\"modelrunner.state.object.ObjectState\"", "__version__": "2"}}, "b": {"__type__": "array", "data": [0, 1, 2], "dtype": "{\"__pickled__\": \"gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAmk4lImIh5RSlChLA4wBPJROTk5K/////0r/\\n////SwB0lGIu\\n\"}", "__attrs__": {"__class__": "\"modelrunner.state.array.ArrayState\"", "__version__": "2"}}}}}
+{"__attrs__": {"__codec__": {"id": "pickle", "protocol": 5}}, "result": {"data": {"__pickled__": "gASVCwEAAAAAAABCBAEAAIAFlfkAAAAAAAAAfZQojAZudW1iZXKUSv////+MBnN0cmluZ5SMBHRl\nc3SUjAdsaXN0XzFklF2UKEsASwFLAmWMB2xpc3RfMmSUXZQoXZQoSwBLAWVdlChLAksDSwRlZYwF\nYXJyYXmUjBJudW1weS5jb3JlLm51bWVyaWOUjAtfZnJvbWJ1ZmZlcpSTlCiWKAAAAAAAAAAAAAAA\nAAAAAAEAAAAAAAAAAgAAAAAAAAADAAAAAAAAAAQAAAAAAAAAlIwFbnVtcHmUjAVkdHlwZZSTlIwC\naTiUiYiHlFKUKEsDjAE8lE5OTkr/////Sv////9LAHSUYksFhZSMAUOUdJRSlHUulC4=\n"}, "__attrs__": {"model": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "format_version": "2", "__type__": "object", "__class__": "modelrunner.results.Result"}}}
