+++ tests/compatibility/2/dict.yaml
@@ -1,30 +1,17 @@
+__attrs__:
+  __codec__:
+    id: pickle
+    protocol: 5
 result:
+  data: !!binary |
+    gAWV+QAAAAAAAAB9lCiMBm51bWJlcpRK/////4wGc3RyaW5nlIwEdGVzdJSMB2xpc3RfMWSUXZQo
+    SwBLAUsCZYwHbGlzdF8yZJRdlChdlChLAEsBZV2UKEsCSwNLBGVljAVhcnJheZSMEm51bXB5LmNv
+    cmUubnVtZXJpY5SMC19mcm9tYnVmZmVylJOUKJYoAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAACAAAA
+    AAAAAAMAAAAAAAAABAAAAAAAAACUjAVudW1weZSMBWR0eXBllJOUjAJpOJSJiIeUUpQoSwOMATyU
+    Tk5OSv////9K/////0sAdJRiSwWFlIwBQ5R0lFKUdS4=
   __attrs__:
-    __model__: '{"class": "MockModel", "name": "model", "description": null, "parameters":
+    model: '{"class": "MockModel", "name": "model", "description": null, "parameters":
       {}}'
-    __version__: '2'
-    __class__: '"modelrunner.results.Result"'
-  state:
-    __attrs__:
-      __keys__: '["a", "b"]'
-      __class__: '"modelrunner.state.dict.DictState"'
-      __version__: '2'
-    a:
-      __type__: array
-      data:
-      - - a
-        - b
-      dtype: '{"__pickled__": "gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAk84lImIh5RSlChLA4wBfJROTk5K/////0r/\n////Sz90lGIu\n"}'
-      __attrs__:
-        __class__: '"modelrunner.state.object.ObjectState"'
-        __version__: '2'
-    b:
-      __type__: array
-      data:
-      - 0
-      - 1
-      - 2
-      dtype: '{"__pickled__": "gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAmk4lImIh5RSlChLA4wBPJROTk5K/////0r/\n////SwB0lGIu\n"}'
-      __attrs__:
-        __class__: '"modelrunner.state.array.ArrayState"'
-        __version__: '2'
+    format_version: '2'
+    __type__: object
+    __class__: modelrunner.results.Result
