+++ tests/compatibility/2/nested.json
@@ -1,0 +1,1 @@
+{"__attrs__": {"__codec__": {"id": "pickle", "protocol": 5}}, "result": {"data": {"__pickled__": "gASVrgAAAAAAAABDqoAFlZ8AAAAAAAAAfZQojAFhlI+UKGgBjAFilJBoA4wSbnVtcHkuY29yZS5u\ndW1lcmljlIwLX2Zyb21idWZmZXKUk5QolhgAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAIAAAAAAAAA\nlIwFbnVtcHmUjAVkdHlwZZSTlIwCaTiUiYiHlFKUKEsDjAE8lE5OTkr/////Sv////9LAHSUYksD\nhZSMAUOUdJRSlHUulC4=\n"}, "__attrs__": {"model": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "format_version": "2", "__type__": "object", "__class__": "modelrunner.results.Result"}}}
