--- tests/compatibility/2/object.json
@@ -1,1 +1,0 @@
-{"result": {"__attrs__": {"__model__": "{\"class\": \"MockModel\", \"name\": \"model\", \"description\": null, \"parameters\": {}}", "__version__": "2", "__class__": "\"modelrunner.results.Result\""}, "state": {"__type__": "array", "data": [{"number": -1, "string": "test", "list_1d": [0, 1, 2], "list_2d": [[0, 1], [2, 3, 4]], "array": [0, 1, 2, 3, 4]}], "dtype": "{\"__pickled__\": \"gASVNwAAAAAAAACMBW51bXB5lIwFZHR5cGWUk5SMAk84lImIh5RSlChLA4wBfJROTk5K/////0r/\\n////Sz90lGIu\\n\"}", "__attrs__": {"__class__": "\"modelrunner.state.object.ObjectState\"", "__version__": "2"}}}}
